+++ packages/react/src/components/NavTabs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The NavTabs component represents the main navigation dock for an application on a mobile device, including up to 5 tabs or buttons for different sections. The component has a fixed position on the bottom of the screen.
 
+## 2023-04-13 `1.1.1`
+
+### 💡 Others
+
+- Added `ariaLabel` property to `NavTabs.Item` subcomponent API. ([#45](https://github.com/TiendaNube/nimbus-patterns/pull/45) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2023-04-04 `1.1.0`
 
 #### 📚 3rd party library updates